requirements.txt
Jinja2==2.8
MarkupSafe==0.23
PyPDF2==1.26.0
PyYAML==3.11
argparse==1.2.1
wsgiref==0.1.2
